+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/data_sources/remote_data_sources/interceptors/analytics_interceptor.dart
@@ -14,6 +14,9 @@ import 'package:dio/dio.dart';
 /// Interceptors can perform a variety of implicit tasks, from authentication
 /// to logging, for every HTTP request/response. Without interception, we will
 /// have to implement these tasks explicitly for each HttpClient method call.
+/// We use analytics interceptor to push notifications, present adds
+/// and collect data for users behaviour.
+/// You should implement one or more methods from the contract.
 class AnalyticsInterceptor extends Interceptor{
 
   @override