+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/repositories/counter_repository.dart
@@ -5,7 +5,7 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import '../../base/count_data_sources/count_data_source.dart';
+import '../../base/data_sources/domain_data_sources/count_data_sources/count_data_source.dart';
 import '../../base/models/count.dart';
 
 ///Can use any type of data source, that follow the contract